move k diffusion out of stable_diffusion
